+++ tests/winrt_tests/angle.test.surfacehost/MockSwapChainPanel.h
@@ -49,7 +49,7 @@ public:
     }
 
     // MockSwapChainPanel
-    int GetSizeChangeRegistrationCount() { return mSizeChangeHandlers.size(); }
+    int GetSizeChangeRegistrationCount() { return static_cast<int>(mSizeChangeHandlers.size()); }
     int GetBoundsQueryCount() { return mBoundsQueryCount; }
     void SetExpectedBounds(ABI::Windows::Foundation::Rect newBounds) { mBounds = newBounds; }
     void UpdateSizeAndSignalChangedEvent(ABI::Windows::Foundation::Rect newBounds)
